noise filtering

+++ GuitarWizardPro.Arduino/ESP32_Host/adc_audio.h
@@ -8,7 +8,7 @@
 namespace Audio::ADC
 {
   const int BUFFER_SIZE = Shared::UNPACKED_PACKET_SIZE;
-  uint16_t* PrimaryBuffer;
+  uint8_t* PrimaryBuffer;
   void Setup(void (*onBufferFullEvent)());
   void Teardown();
 }
